Annotate pmprintf and __pmNotifyErr with printf format.
If the build compiler is gcc, we now annotate the pmprintf
and __pmNotifyErr routines so that callers are informed if
using them inappropriately (incorrect types, wrong argument
count, that sort of thing).

Fixed up the newly introduced warnings - a couple of which
were indeed bugs.

+++ src/libpcp/src/util.c
@@ -561,7 +561,8 @@ void
 __pmNoMem(const char *where, size_t size, int fatal)
 {
     __pmNotifyErr(fatal ? LOG_ERR : LOG_WARNING,
-		 "%s: malloc(%d) failed: %s", where, size, strerror(errno));
+			"%s: malloc(%d) failed: %s",
+			where, (int)size, strerror(errno));
     if (fatal)
 	exit(1);
 }
